fix(network): cast entity generation to uint16_t in datalink serialization
MSVC (/WX) fails Build (windows-latest) on WorldBroadcaster.cpp:2656 and :2677
with C4244: the datalink track/threat records assign a uint32_t EntityId
generation into their uint16_t wire fields (targetGen / emitterGen). Add the
explicit narrowing cast, matching the existing static_cast<uint16_t> on the
threatCount field just above. The generation always fits in 16 bits (the wire
gen field is uint16_t by design), so this is lossless in practice.

Unblocks the Windows build for #498 (Epic F).

+++ engine/net/WorldBroadcaster.cpp
@@ -2653,7 +2653,7 @@ void WorldBroadcaster::broadcastDatalink(uint64_t tickIndex) {
         for (const sensor::FusedTrack& t : tracks) {
             DatalinkTrack r{};
             r.targetIdx = t.id.index;
-            r.targetGen = t.id.generation;
+            r.targetGen = static_cast<uint16_t>(t.id.generation);
             r.typeIndex = t.typeIndex;
             r.factionIndex = t.factionIndex;
             r.state = static_cast<uint8_t>(t.state);
@@ -2674,7 +2674,7 @@ void WorldBroadcaster::broadcastDatalink(uint64_t tickIndex) {
             const sensor::ThreatWarning& w = threats->threats[i];
             DatalinkThreat r{};
             r.emitterIdx = w.emitterId.index;
-            r.emitterGen = w.emitterId.generation;
+            r.emitterGen = static_cast<uint16_t>(w.emitterId.generation);
             r.emitterTypeIndex = w.emitterTypeIndex;
             r.emitterFactionIndex = w.emitterFactionIndex;
             r.channel = static_cast<uint8_t>(w.channel);
